move managers to its own dir and exclude them from coverage
