+++ packages/dashboard/src/managers/door-state-manager.ts
