+++ packages/dashboard/src/managers/emergency-manager.ts
