+++ packages/dashboard/src/managers/robot-trajectory-manager.ts
